+++ src/tailscale_manager/models/__init__.py
@@ -1,9 +1,11 @@
 from tailscale_manager.models.acl import AclConfig
+from tailscale_manager.models.agenix_sync import AgenixSyncResult
 from tailscale_manager.models.auth_key import TailscaleAuthKey
 from tailscale_manager.models.device import TailscaleDevice
 
 __all__ = [
     "AclConfig",
+    "AgenixSyncResult",
     "TailscaleAuthKey",
     "TailscaleDevice",
 ]